django/requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20230430
Django==5.1.1
django-celery-beat==2.7.0
django-import-export==4.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
frozendict==2.4.4
html5lib==1.1
idna==3.8
inflection==0.5.1
kombu==5.4.1
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.3.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.43
